refactor(api): restructure src/ into domain-based modules with shared helpers
Split flat services/ + lib/ + utils/ + types/ into domain-based folders:
articles/, projects/, offerings/, snippets/, experiences/, comments/,
guestbook/, stats/, search/, users/, shared/, storage/, http/.

Extracted shared helpers:
- errors.ts (reportError), db.ts (DbClient + AdminContext)
- draft-access.ts (assertPublishedOrAdmin)
- image-lifecycle.ts (handleImageUpload)
- markdown-rendering.ts (buildContentRendering)
- pagination.ts, validation.ts

Fixed package.json types path and added ./env export condition.
