+++ packages/api/package.json
@@ -5,10 +5,13 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
+      "types": "./dist/src/index.d.ts",
       "default": "./src/index.ts"
     },
-    "./env": "./env.ts"
+    "./env": {
+      "types": "./dist/env.d.ts",
+      "default": "./env.ts"
+    }
   },
   "scripts": {
     "build": "tsc",
