+++ packages/api/src/index.ts
@@ -1,22 +1,22 @@
 /** biome-ignore-all lint/performance/noBarrelFile: This is a barrel file */
 
-export type { PaginatedResult, PaginationInput } from './lib/base-service';
-export { handleImageUpdate, handleImageUpload } from './lib/base-service';
-export { escapeSearchTerm, isValidBase64, validateSearchQuery } from './lib/validation';
-export {
-  blogService,
-  commentService,
-  experienceService,
-  guestbookService,
-  projectService,
-  searchService,
-  serviceService,
-  snippetService,
-  statsService,
-  userService,
-} from './services';
-export { deleteFile, uploadImage } from './storage';
-
+export * as articleEngagementService from './articles/article-engagement.service';
+export * as articleRecommendationsService from './articles/article-recommendations.service';
+export * as articlesService from './articles/articles.service';
+export * as commentsService from './comments/comments.service';
+export * as experiencesService from './experiences/experiences.service';
+export * as guestbookService from './guestbook/guestbook.service';
 // Error handling
-export type { ApiErrorResponse, ApiSuccessResponse } from './types/error-response';
-export { createSuccessResponse, handleApiError } from './utils/error-handler';
+export type { ApiErrorResponse, ApiSuccessResponse } from './http/error-response';
+export { createSuccessResponse, handleApiError } from './http/response';
+export * as offeringService from './offerings/offerings.service';
+export * as offeringsService from './offerings/offerings.service';
+export * as projectsService from './projects/projects.service';
+export * as searchService from './search/search.service';
+export { handleImageUpload } from './shared/image-lifecycle';
+export type { PaginatedResult, PaginationInput } from './shared/pagination';
+export { escapeSearchTerm, isValidBase64, validateSearchQuery } from './shared/validation';
+export * as snippetsService from './snippets/snippets.service';
+export * as statsService from './stats/stats.service';
+export { deleteFile, uploadImage } from './storage/blob-storage';
+export * as usersService from './users/users.service';
