+++ packages/api/src/search/search.service.ts
@@ -1,12 +1,9 @@
-// biome-ignore lint/performance/noNamespaceImport: Sentry SDK requires namespace import
-import * as Sentry from '@sentry/node';
-import type { db as DB } from '@xbrk/db/client';
 import { articles, project, snippet } from '@xbrk/db/schema';
 import { ValidationError } from '@xbrk/errors';
 import { and, eq, ilike, or } from 'drizzle-orm';
-import { escapeSearchTerm, validateSearchQuery } from '../lib/validation';
-
-type DbClient = typeof DB;
+import type { DbClient } from '../shared/db';
+import { reportError } from '../shared/errors';
+import { escapeSearchTerm, validateSearchQuery } from '../shared/validation';
 
 interface SearchResult {
   description: string | null;
@@ -21,15 +18,8 @@ interface SearchResults {
   snippets: SearchResult[];
 }
 
-/**
- * Searches across articles, projects, and snippets using case-insensitive ILIKE.
- * Only published (non-draft) records are included. Returns up to 5 results per entity type.
- * Special SQL characters (`%`, `_`, `\`) in the query are escaped before matching.
- * @throws {Error} If query validation fails
- */
 export async function query(db: DbClient, input: { query: string }): Promise<SearchResults> {
   try {
-    // Validate search query
     const validation = validateSearchQuery(input.query);
     if (!validation.valid) {
       throw new ValidationError(validation.error ?? 'Invalid search query');
@@ -88,8 +78,7 @@ export async function query(db: DbClient, input: { query: string }): Promise<Sea
     if (error instanceof ValidationError) {
       throw error;
     }
-    Sentry.captureException(error);
-    console.error('[search.query] Database error:', error);
+    reportError('search.query', error);
     return {
       articles: [],
       projects: [],
