--- packages/api/src/services/index.ts
@@ -1,10 +1,0 @@
-export * as blogService from './blog';
-export * as commentService from './comment';
-export * as experienceService from './experience';
-export * as guestbookService from './guestbook';
-export * as projectService from './project';
-export * as searchService from './search';
-export * as serviceService from './service';
-export * as snippetService from './snippet';
-export * as statsService from './stats';
-export * as userService from './user';
