+++ packages/api/src/shared/db.ts
@@ -1,0 +1,7 @@
+import type { db as DB } from '@xbrk/db/client';
+
+export type DbClient = typeof DB;
+
+export interface AdminContext {
+  user: { id: string; role: 'admin' };
+}
