+++ packages/api/src/shared/errors.ts
@@ -1,0 +1,7 @@
+// biome-ignore lint/performance/noNamespaceImport: Sentry SDK requires namespace import
+import * as Sentry from '@sentry/node';
+
+export function reportError(context: string, error: unknown): void {
+  Sentry.captureException(error);
+  console.error(`[${context}]`, error);
+}
