+++ packages/api/src/shared/pagination.ts
@@ -1,0 +1,14 @@
+export interface PaginationInput {
+  limit?: number;
+  page?: number;
+}
+
+export interface PaginatedResult<T> {
+  data: T[];
+  pagination: {
+    page: number;
+    limit: number;
+    total: number;
+    totalPages: number;
+  };
+}
