+++ packages/api/src/snippets/snippets.service.ts
@@ -1,29 +1,26 @@
-// biome-ignore lint/performance/noNamespaceImport: Sentry SDK requires namespace import
-import * as Sentry from '@sentry/node';
 import { CreateSnippetSchema, UpdateSnippetSchema } from '@xbrk/db/api-schemas';
 import type { db as DB } from '@xbrk/db/client';
 import { snippet } from '@xbrk/db/schema';
 import { InternalServerError, NotFoundError } from '@xbrk/errors';
-import { markdownToHastJson, RENDERING_VERSION } from '@xbrk/md/processor';
 import { desc, eq, sql } from 'drizzle-orm';
 import type { z } from 'zod/v4';
+import { assertPublishedOrAdmin } from '../shared/draft-access';
+import { reportError } from '../shared/errors';
+import { buildContentRendering } from '../shared/markdown-rendering';
 
 type DbClient = typeof DB;
 
-/** Returns all snippets including drafts. For admin use only. */
 export async function getAll(db: DbClient) {
   try {
     return await db.query.snippet.findMany({
       orderBy: desc(snippet.id),
     });
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.getAll] Database error:', error);
+    reportError('snippets.getAll', error);
     return [];
   }
 }
 
-/** Returns only published (non-draft) snippets. */
 export async function getAllPublic(db: DbClient) {
   try {
     return await db.query.snippet.findMany({
@@ -31,13 +28,11 @@ export async function getAllPublic(db: DbClient) {
       where: eq(snippet.isDraft, false),
     });
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.getAllPublic] Database error:', error);
+    reportError('snippets.getAllPublic', error);
     return [];
   }
 }
 
-/** Returns a single snippet by ID. */
 export async function getById(db: DbClient, input: { id: string }) {
   try {
     const query = db.query.snippet
@@ -47,17 +42,11 @@ export async function getById(db: DbClient, input: { id: string }) {
       .prepare('get_snippet_by_id');
     return await query.execute({ id: input.id });
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.getById] Database error:', error);
+    reportError('snippets.getById', error);
     return undefined;
   }
 }
 
-/**
- * Returns a single snippet by slug.
- * Draft snippets are only accessible to admins.
- * @throws {Error} If snippet not found or is a draft and requester is not admin.
- */
 export async function getBySlug(db: DbClient, input: { slug: string }, session?: { user: { role: string } } | null) {
   try {
     const query = db.query.snippet
@@ -67,37 +56,29 @@ export async function getBySlug(db: DbClient, input: { slug: string }, session?:
       .prepare('get_snippet_by_slug');
     const result = await query.execute({ slug: input.slug });
 
-    if (!result) {
-      throw new NotFoundError('Snippet not found');
-    }
-
-    if (result.isDraft && session?.user.role !== 'admin') {
-      throw new NotFoundError('Snippet is not public');
-    }
+    assertPublishedOrAdmin(result, 'Snippet', session);
 
     return result;
   } catch (error) {
     if (error instanceof NotFoundError) {
       throw error;
     }
-    Sentry.captureException(error);
-    console.error('[snippet.getBySlug] Database error:', error);
+    reportError('snippets.getBySlug', error);
     throw new InternalServerError('Failed to fetch snippet');
   }
 }
 
 export async function create(db: DbClient, input: z.infer<typeof CreateSnippetSchema>) {
   try {
-    const contentRendering = input.code ? await markdownToHastJson(input.code) : null;
+    const { contentRendering, contentRenderingVersion } = await buildContentRendering(input.code);
 
     await db.insert(snippet).values({
       ...input,
       contentRendering,
-      contentRenderingVersion: RENDERING_VERSION,
+      contentRenderingVersion,
     });
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.create] Database error:', error);
+    reportError('snippets.create', error);
     throw new InternalServerError('Failed to create snippet');
   }
 }
@@ -105,27 +86,48 @@ export async function create(db: DbClient, input: z.infer<typeof CreateSnippetSc
 export async function update(db: DbClient, input: z.infer<typeof UpdateSnippetSchema>) {
   try {
     const { id, ...updateData } = input;
+
+    const existing = await db.query.snippet.findFirst({
+      where: eq(snippet.id, id),
+    });
+    if (!existing) {
+      throw new NotFoundError('Snippet not found');
+    }
+
     const setData: Record<string, unknown> = { ...updateData };
 
     if (input.code !== undefined) {
-      setData.contentRendering = input.code ? await markdownToHastJson(input.code) : null;
-      setData.contentRenderingVersion = RENDERING_VERSION;
+      const { contentRendering, contentRenderingVersion: version } = await buildContentRendering(input.code);
+      setData.contentRendering = contentRendering;
+      setData.contentRenderingVersion = version;
     }
 
     await db.update(snippet).set(setData).where(eq(snippet.id, id));
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.update] Database error:', error);
+    if (error instanceof NotFoundError) {
+      throw error;
+    }
+    reportError('snippets.update', error);
     throw new InternalServerError('Failed to update snippet');
   }
 }
 
 export async function remove(db: DbClient, id: string) {
   try {
+    const existing = await db.query.snippet.findFirst({
+      where: eq(snippet.id, id),
+    });
+
+    if (!existing) {
+      throw new NotFoundError('Snippet not found');
+    }
+
     await db.delete(snippet).where(eq(snippet.id, id));
   } catch (error) {
-    Sentry.captureException(error);
-    console.error('[snippet.remove] Database error:', error);
+    if (error instanceof NotFoundError) {
+      throw error;
+    }
+    reportError('snippets.remove', error);
     throw new InternalServerError('Failed to delete snippet');
   }
 }
