+++ packages/api/src/storage/blob-storage.ts
@@ -1,21 +1,14 @@
 // biome-ignore lint/performance/noNamespaceImport: Sentry SDK requires namespace import
 import * as Sentry from '@sentry/node';
 import { del, put } from '@vercel/blob';
-import { InternalServerError, ValidationError } from '@xbrk/errors';
+import { HttpError, InternalServerError, ValidationError } from '@xbrk/errors';
 import { generateSlug } from '@xbrk/utils';
-import { isValidBase64 } from './lib/validation';
+import { isValidBase64 } from '../shared/validation';
 
-export const MAX_IMAGE_SIZE = 5 * 1024 * 1024; // 5MB
-const allowedDomains = ['vercel-blob.com', 'blob.vercel-storage.com'];
+export const MAX_IMAGE_SIZE = 5 * 1024 * 1024;
 
-/**
- * Uploads an image to Vercel Blob storage.
- * @param folder - The folder path in blob storage
- * @param image - Base64 encoded image string
- * @param slug - Slug for the filename (will be sanitized)
- * @returns The public URL of the uploaded image
- * @throws {Error} If validation fails or upload fails
- */
+const allowedDomains = new Set(['vercel-blob.com', 'blob.vercel-storage.com']);
+
 export async function uploadImage(folder: string, image: string, slug: string): Promise<string> {
   try {
     if (!image?.trim()) {
@@ -33,7 +26,6 @@ export async function uploadImage(folder: string, image: string, slug: string): 
       );
     }
 
-    // Sanitize slug to ensure it's URL-safe
     const safeSlug = generateSlug(slug);
     const fileName = `${safeSlug}-${Date.now()}.avif`;
     const imageBuffer = Buffer.from(image, 'base64');
@@ -45,16 +37,14 @@ export async function uploadImage(folder: string, image: string, slug: string): 
     });
     return url;
   } catch (_error) {
+    if (_error instanceof HttpError) {
+      throw _error;
+    }
     Sentry.captureException(_error);
     throw new InternalServerError('Failed to upload image', { cause: _error as Error });
   }
 }
 
-/**
- * Deletes a file from Vercel Blob storage.
- * @param url - The URL of the file to delete
- * @throws {Error} If validation fails or deletion fails
- */
 export async function deleteFile(url: string): Promise<void> {
   try {
     if (!url?.trim()) {
@@ -62,12 +52,15 @@ export async function deleteFile(url: string): Promise<void> {
     }
 
     const urlObj = new URL(url);
-    if (!allowedDomains.some((domain) => urlObj.hostname.includes(domain))) {
+    if (!allowedDomains.has(urlObj.hostname)) {
       throw new ValidationError('URL does not belong to an allowed storage domain');
     }
 
     await del(url);
   } catch (_error) {
+    if (_error instanceof HttpError) {
+      throw _error;
+    }
     Sentry.captureException(_error);
     throw new InternalServerError('Failed to delete file', { cause: _error as Error });
   }
